apriltag_vision_py/requirements.txt
opencv-python
numpy
apriltag
pynetworktables